GROOT START!
Forked `orgit` repo (GPL3) in order to use it to figure out how to
make `groot` (GPL3).

`orgit` is a new kind of link for `org-mode`.
`groot` is a new kind of link for `org-mode`.

So hopefully `orgit` can provide a lot of info about how the hell to
do this...

First though, put aside all the CI/CD and packing up for MELPA type
stuff.