Fix bug where error did not allow link to be stored
`groot` should never stop `org-link` from functioning.

However, the "downgrade errors to warnings" was causing links to not
get stored.

Bug was that a value was being returned when the warning was emitted,
which confused `org-link` into not storing a link.

Fix bug by explicitly returning nil in ~groot-link--store~ in
error/warning cases.